requirements.txt
transformers==4.49.0
datasets
torchmetrics
torchvision
